refactor: corrige inconsistências após um merge.
